+++ class/wallets/abstract-hd-electrum-wallet.ts
@@ -18,7 +18,7 @@ import { randomBytes } from '../rng';
 import { AbstractHDWallet } from './abstract-hd-wallet';
 import { CreateTransactionResult, CreateTransactionTarget, CreateTransactionUtxo, Transaction, Utxo } from './types';
 import { SilentPayment, UTXOType as SPUTXOType, UTXO as SPUTXO } from 'silent-payments';
-import { isValidBech32Address } from '../../utils/isValidBech32Address';
+import { isValidBech32Address } from '../../util/isValidBech32Address.ts';
 
 const ECPair = ECPairFactory(ecc);
 const bip32 = BIP32Factory(ecc);
