+++ screen/send/psbtMultisig.tsx
@@ -26,7 +26,7 @@ import loc from '../../loc';
 import { BitcoinUnit } from '../../models/bitcoinUnits';
 import { useStorage } from '../../hooks/context/useStorage';
 import { useExtendedNavigation } from '../../hooks/useExtendedNavigation';
-import { combinePSBTs } from '../../utils/combinePSBTs';
+import { combinePSBTs } from '../../util/combinePSBTs.ts';
 import { MultisigHDWallet } from '../../class';
 import assert from 'assert';
